[fix] A função de deletar imagem não ocntava com o caso de chegar null ao invés de uma instância de App/Image

+++ app/Domains/Image/Repository/ImageRepository.php
@@ -27,11 +27,13 @@ class ImageRepository extends LogicsAndRepositories
 	}
 
 	/**
-	 * @param Image $image
+	 * @param Image|null $image
 	 * @return void
 	 */
-	public function deleteImage(Image $image){
-		Storage::delete($image->getPath()); // TODO: quem sabe adicionar um failproof aqui
-		$image->delete();
+	public function deleteImage(?Image $image){
+		if(!is_null($image)){
+			Storage::delete($image->getPath()); // TODO: quem sabe adicionar um failproof aqui
+			$image->delete();
+		}
 	}
 }
